refactor: moved test to new file, refactor and documentation
- Move test relating to ClipBoardManager.java to a separate file.
- Refactored tests in the new ClipBoardManagerTest file
- Added documentation to tests
